pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "antennalab"
version = "0.1.0"
description = "Laptop-based antenna/RF toolkit"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
antennalab = "antennalab.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
